fixed compiler warnings

+++ src/marshalling.cpp
@@ -546,7 +546,7 @@ TEST(unmarshall_level_one_simple,
 
 TEST(unmarshall_level_one_complex,
      static const struct S { double d; char c; int64_t i; } s1 = { 2.5, 'C',
-         0xfffffffffffffffeLL }, s2 = { -2.5, 'D', 50LL }, szero { 0, 0, 0LL };
+	(int64_t) 0xfffffffffffffffeLL }, s2 = { -2.5, 'D', 50LL }, szero { 0, 0, 0LL };
      static const S * ARGS[] { 0, &s1, 0, &s1, &s2, 0, &s1 };
      static const int PTR_ARRAY[] =
      {
